remove unused tables
